refactor: externalise the 'report' features to another file
