+++ tests/test_misc.py
@@ -27,12 +27,6 @@ import pathlib
 import numpy as np
 import pytest
 
-from conftest import has_dependency, requires
-
-if has_dependency("pyvista"):
-    from pyvista.plotting import system_supports_plotting
-
-from ansys.mapdl import core as pymapdl
 from ansys.mapdl.core.misc import (
     check_valid_ip,
     check_valid_port,
@@ -44,18 +38,7 @@ from ansys.mapdl.core.misc import (
     requires_package,
     run_as_prep7,
 )
-
-
-@requires("pyvista")
-def test_report():
-    report = pymapdl.Report(
-        additional=["matplotlib", "pyvista", "pyiges", "tqdm"],
-        gpu=system_supports_plotting(),
-    )
-    assert "PyAnsys Software and Environment Report" in str(report)
-
-    # Check that when adding additional (repeated) packages, they appear only once
-    assert str(report).count("pyvista") == 1
+from conftest import requires
 
 
 @pytest.mark.parametrize(
@@ -232,68 +215,6 @@ def test_load_file_local(mapdl, tmpdir, file_):
 
     mapdl.slashdelete(file_)
     assert file_ not in mapdl.list_files()
-
-
-def test_plain_report():
-    from ansys.mapdl.core.misc import Plain_Report
-
-    core = ["numpy", "ansys.mapdl.reader"]
-    optional = ["pyvista", "tqdm"]
-    additional = ["scipy", "ger"]
-
-    report = Plain_Report(core=core, optional=optional, additional=additional, gpu=True)
-    rep_str = report.__repr__()
-
-    for each in core + optional + additional:
-        assert each in rep_str
-
-    # There should be only one package not found ("ger")
-    assert "Package not found" in rep_str
-    not_found_packages = 1
-
-    # Plus the not additional packages
-    if not has_dependency("pyvista"):
-        not_found_packages += 1
-    if not has_dependency("tqdm"):
-        not_found_packages += 1
-    if not has_dependency("ansys.mapdl.reader"):
-        not_found_packages += 1
-    if not has_dependency("scipy"):
-        not_found_packages += 1
-    if not has_dependency("pexpect"):
-        not_found_packages += 1
-
-    _rep_str = rep_str.replace("Package not found", "", not_found_packages)
-    assert "Package not found" not in _rep_str
-
-    assert "\n" in rep_str
-    assert len(rep_str.splitlines()) > 3
-
-    assert "Core packages" in rep_str
-    assert "Optional packages" in rep_str
-    assert "Additional packages" in rep_str
-
-    # Plain report should not represent GPU details evenif asked for
-    assert "GPU Details" not in rep_str
-
-
-def test_plain_report_no_options():
-    from ansys.mapdl.core.misc import Plain_Report
-
-    core = ["numpy", "ansys.mapdl.reader"]
-
-    report = Plain_Report(core=core)
-    rep_str = report.__repr__()
-
-    for each in core:
-        assert each in rep_str
-
-    assert "\n" in rep_str
-    assert len(rep_str.splitlines()) > 3
-
-    assert "Core packages" in rep_str
-    assert "Optional packages" not in rep_str
-    assert "Additional packages" not in rep_str
 
 
 def test_requires_package_decorator():
